pyproject.toml
[project]
name = "dev-report"
version = "0.1.0"
description = "Daily/weekly/monthly dev activity reports across git repositories"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0",
    "gitpython>=3.1",
]

[project.scripts]
dev-report = "dev_report.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dev_report"]
